requirements.txt
Django==2.0.7
pytz==2018.5
djangorestframework==3.8.2
Markdown==2.6.11
django-filter==1.0.4
kazoo==2.5.0
six==1.11.0
redis==2.10.6
